% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LukaFunctions.R
\name{get_authentication_token}
\alias{get_authentication_token}
\title{Get an API authentication token. The function defines the global variable, auth_token,  which is a default value in all API functions. The user only needs to run this function successfully to use the package, no assignment required. If a token expires simply rerun the function to generate a new one.}
\usage{
get_authentication_token(CLIENT_ID, CLIENT_SECRET)
}
\arguments{
\item{CLIENT_ID}{A Spotify client ID which the user needs to acquire prior to using the package.}

\item{CLIENT_SECRET}{A Spotify client secret key which the user needs to acquire prior to using the package.}
}
\value{
auth_token
}
\description{
Get an API authentication token. The function defines the global variable, auth_token,  which is a default value in all API functions. The user only needs to run this function successfully to use the package, no assignment required. If a token expires simply rerun the function to generate a new one.
}
\examples{
get_authentication_token(CLIENT_ID, CLIENT_SECRET)
}
